requirements.txt
fastapi
uvicorn[standard]
google-auth-oauthlib
google-api-python-client
python-dotenv
aiofiles
